maxbrain/agents/calendar.py
from google.ai import generativelanguage as glm

# Skill Agent Implementation (Calendar Example)
calendar_agent = glm.Tool(
    name="ScheduleMaster",
    description="Manages calendar events",
    function_declarations=[
        glm.FunctionDeclaration(
            name="create_event",
            description="Creates calendar events",
            parameters={
                "type_": "object",
                "properties": {
                    "title": {"type_": "string"},
                    "date": {"type_": "string", "format": "date-time"},
                    "duration": {"type_": "integer"},
                    "attendees": {"type_": "array", "items": {"type_": "string"}}
                },
                "required": ["title", "date"]
            }
        )
    ]
)

def create_event(title, date, duration=60, attendees=None):
    """
    Creates a calendar event with the specified parameters.
    
    Args:
        title (str): The title of the event
        date (str): The date and time of the event in ISO format
        duration (int, optional): The duration of the event in minutes. Defaults to 60.
        attendees (list, optional): List of attendee email addresses. Defaults to None.
    
    Returns:
        dict: The created event details
    """
    # Implementation would connect to calendar API
    event = {
        "title": title,
        "date": date,
        "duration": duration,
        "attendees": attendees or []
    }
    
    # Here you would add code to actually create the event in a calendar system
    
    return {
        "status": "success",
        "message": f"Event '{title}' created successfully",
        "event": event
    }